refactored to GitDirLocator
